pyproject.toml
[tool.poetry]
name = "projet-10"
version = "0.1.0"
description = ""
authors = ["LeJinge <jeremycarmona28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
